Created HTTP request Filter API: apiFilters.js
As of yet, the API has only been extensively applied to the tossController;
but that is the most important collection to filter and sort. The routing
established here should allow for displaying active Toss's on the homepage,
searching/sorting past Toss's (Toss's with a currentPhase value of 5), as well
as displaying Toss's that start the next day.

This filter API can be readily applied to other types of Collections/Documents
in MongoDB.